Introduces ServerEventDistributor so ZeroC Ice and the plugin host are
both EventSubscribers, and brings in the live-doc rewrite.

Conflict resolution highlights:
- user_id widened i32->i64 end-to-end (ClientInfo/FFI/JWT/SessionInfo)
- file-server: dropped me.rs in favor of unify's superset private.rs;
  unioned SessionClaims (scope + uid + reg) and kept the durable
  scope-based per-user storage namespacing (survives cert regeneration)
- documents: unioned owner-tracking + shared-with ACL
- live-doc: took the unify rewrite wholesale (state/lib/persistence)

Verified: cargo test (file-server 107, live-doc 26, host 15) and a full
dev_build (C++ server + plugins, ice=ON, LTO) all green.

+++ 3rdparty/mumble-plugin-host/host/include/mumble_plugin_host.h
@@ -183,6 +183,9 @@ extern "C" {
 /**
  * Notify the host of a newly connected client.
  *
+ * `user_id` is the registered Mumble user id, or `-1` for an
+ * unregistered guest.
+ *
  * # Safety
  * `handle` must be valid; `username` and `cert_hash` must be NUL-terminated
  * UTF-8 strings (or NULL, which is treated as the empty string).  `user_id`
@@ -194,7 +197,7 @@ void plugin_host_on_client_connected(struct PluginHostHandle *handle,
                                      uint32_t session,
                                      const char *username,
                                      const char *cert_hash,
-                                     int32_t user_id);
+                                     int64_t user_id);
 
 /**
  * Notify the host of a disconnected client.
